fix(ui): keep routing persistence out of browser bundles (#20594)
* fix(ui): keep routing persistence out of browser bundles

* refactor(shared): split browser-safe routing policy

---------
