+++ packages/shared/src/local-inference/routing-preferences.ts
@@ -14,64 +14,23 @@ import { hostname } from "node:os";
 import path from "node:path";
 import { localInferenceRoot } from "./paths.js";
 import {
+  isRoutingPolicy,
+  type RoutingPolicy,
+  type RoutingPreferences,
+} from "./routing-policy.js";
+import {
   AGENT_MODEL_SLOTS,
   type AgentModelSlot,
   TEXT_GENERATION_SLOTS,
 } from "./types.js";
 
-export type RoutingPolicy =
-  | "manual"
-  | "auto"
-  | "local-only"
-  | "cloud-only"
-  | "cheapest"
-  | "fastest"
-  | "prefer-local"
-  | "round-robin";
-
-/**
- * The full set of selectable policies, in display order. Kept as a runtime
- * value (not just a type) so route-layer validation and the settings UI share
- * one source of truth for "which policies are accepted".
- *
- * `local-only` / `cloud-only` are the canonical per-slot replacements for the
- * global `ELIZA_LOCAL_ONLY` env hack: `local-only` is a hard guarantee (never
- * falls through to cloud), `cloud-only` never dispatches on-device.
- */
-export const ROUTING_POLICIES: readonly RoutingPolicy[] = [
-  "manual",
-  "auto",
-  "local-only",
-  "cloud-only",
-  "cheapest",
-  "fastest",
-  "prefer-local",
-  "round-robin",
-] as const;
-
-export function isRoutingPolicy(value: unknown): value is RoutingPolicy {
-  return (
-    typeof value === "string" &&
-    (ROUTING_POLICIES as readonly string[]).includes(value)
-  );
-}
-
-export const DEFAULT_ROUTING_POLICY: RoutingPolicy = "prefer-local";
-
-export interface RoutingPreferences {
-  /**
-   * Explicit provider override per agent slot. Empty record = no overrides,
-   * runtime picks the highest-priority registered handler.
-   */
-  preferredProvider: Partial<Record<AgentModelSlot, string>>;
-  /**
-   * Per-slot policy. "manual" honours `preferredProvider` verbatim;
-   * everything else lets the router-handler compute a winner from the
-   * policy rule set. Absent = "prefer-local" so local models and
-   * subscriptions are tried before direct paid APIs and managed cloud.
-   */
-  policy: Partial<Record<AgentModelSlot, RoutingPolicy>>;
-}
+export {
+  DEFAULT_ROUTING_POLICY,
+  isRoutingPolicy,
+  ROUTING_POLICIES,
+  type RoutingPolicy,
+  type RoutingPreferences,
+} from "./routing-policy.js";
 
 interface RoutingFile {
   version: 1;
